Add approximate queue position to system.runtime.queries
The approximate position is evaluated by cycling through
the resource groups which contain queued queries.

Ex. Resource group a.b has queries q1, q2, q3
    Resource group a.c has queries q4, q5, q6
    Resource group b.a has queries q7, q8
    Resource group b.b has queries q9, q10

    Order is q1, q4, q7, q9, q2, q5, q8, q10, q3, q6

It is not possible to come up with an actual order without knowing how long each query will take.
This approximation is one possible ordering and provides some notion of the position in the queue.